updated backend to allow for manual shortlink deletion :^)

+++ pages/api/delete/[slug].js
@@ -1,8 +1,10 @@
 import { prisma } from "../../../utils/prisma";
+import bcrypt from "bcrypt";
 
 export default async (req, res) => {
-  // TODO: Replace Search Slug with try catch delete https://www.prisma.io/docs/reference/api-reference/prisma-client-reference#delete
-  const searchSlug = await prisma.shortLink.findUnique({
+
+  // gets the slug's hashedPin
+  const searchSlug = await prisma.shortLink.findUnique({ 
     where: {
         slug: req.query.slug
     }
@@ -12,13 +14,28 @@ export default async (req, res) => {
     return res.status(400).json({ message: `a slug called ${req.query.slug} does not exist :(` });
   }
 
-  const deletedSlug = await prisma.shortLink.delete({
-    where: {
-        slug: req.query.slug
-    }
-  })
+  if (!req.body.pin) {
+    return res.status(400).json({ message: `you must provide the currect pin to delete your slug >:()` });
+  }
 
-  return res.status(200).json({
-    message: `${deletedSlug.slug} has been deleted. farewell little guy.`,
+  const validPin = bcrypt.compare(req.body.pin, searchSlug.hashedPin, async (err, data) => {
+    if (err) {
+      return err
+    }
+  
+    if (data) {
+      const deletedSlug = await prisma.shortLink.delete({
+        where: {
+            slug: req.query.slug
+        }
+      })
+    
+      return res.status(200).json({
+        message: `${deletedSlug.slug} has been deleted. farewell little guy.`,
+      });
+    } else {
+      return res.status(401).json({ message: "incorrect pin >:(" })
+    }
   });
+
 };
